refactor: decompose and extract fast-actions to a dedicated hook
